backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
httpx>=0.25.0
python-multipart>=0.0.6

# Database drivers
# SQLite is built into Python, no driver needed
# psycopg2-binary==2.9.9  # Uncomment if using PostgreSQL

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
